requirements.txt
numpy
matplotlib
gradio==5.23.0
pillow
